requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dj-database-url==3.1.0
Django==4.2.24
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
multidict==6.7.1
numpy==2.1.1
opencv-python==4.10.0.84
packaging==26.0
pillow==11.1.0
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
python-dotenv==1.2.1
razorpay==2.0.0
requests==2.32.3
setuptools==82.0.0
sqlparse==0.5.5
twilio==9.10.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.30.6
wheel==0.46.3
whitenoise==6.11.0
yarl==1.22.0
